[front] refactor: change router and directory name
- modified
`/ranks` to `/ranking` when it comes to route client url.
changed `RankPage/` to `RankingPage/` to unify router and directory name.
