+++ client/src/pages/RankingPage/AreaSelect.tsx
